# src/main/java/com/example/demo/model/User.java
package com.example.demo.model;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.EnumType;
import jakarta.persistence.Enumerated;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import java.util.Objects;

@Entity
@Table(name = "users")
public class User {
  @Id
  @GeneratedValue(strategy = GenerationType.IDENTITY)
  private long id;

  @Column(nullable = false, unique = true)
  private String email;

  @Column(nullable = false)
  private String password;

  private String name;

  @Enumerated(EnumType.STRING)
  private Role role = Role.USER;

  public enum Role {
    USER,
    ADMIN
  };

  public User() {}

  public User(long id, String email, String password, String name, Role role) {
    this.id = id;
    this.email = email;
    this.password = password;
    this.name = name;
    this.role = role;
  }

  public long getId() {
    return id;
  }

  public void setId(long id) {
    this.id = id;
  }

  public String getEmail() {
    return email;
  }

  public void setEmail(String email) {
    this.email = email;
  }

  public String getPassword() {
    return password;
  }

  public void setPassword(String password) {
    this.password = password;
  }

  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public Role getRole() {
    return role;
  }

  public void setRole(Role role) {
    this.role = role;
  }

  @Override
  public boolean equals(Object obj) {
    if (this == obj) {
      return true;
    }
    if (!(obj instanceof User)) {
      return false;
    }
    User other = (User) obj;
    return other.canEqual(this)
        && id == other.id
        && Objects.equals(email, other.email)
        && Objects.equals(password, other.password)
        && Objects.equals(name, other.name)
        && Objects.equals(role, other.role);
  }

  protected boolean canEqual(Object other) {
    return other instanceof User;
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, email, password, name, role);
  }

  @Override
  public String toString() {
    return "User(id="
        + id
        + ", email="
        + email
        + ", password=[REDACTED]"
        + ", name="
        + name
        + ", role="
        + role
        + ")";
  }
}